+++ src/Image.h
@@ -7,10 +7,10 @@
 #include "Structs.h"
 
 //loads a bmp file into memory for compression
-int LoadBMP(char *fname, unsigned char *&header_addr, int &header_size, double *&img_addr, int &img_size);
+int LoadBMP(char *fname, unsigned char *&header_addr, int &header_size, float *&img_addr, int &img_size);
 
 //saves a bmp file after decompression
-int SaveBMP(char *fname, unsigned char *header_data, int header_size, double *img_data, int img_size);
+int SaveBMP(char *fname, unsigned char *header_data, int header_size, float *img_data, int img_size);
 
 //saves a compressed file (extension is .wlt)
 int SaveWLT(char *fname, wlt_header_info wlt, unsigned char *bmp_header_data,
